Fix name typo(no content change)
Hi Tianpei, 
I just realized there is a typo with my last name when I was looking at the compiled survey paper, I did not find it when I was looking at the github markdown page. This pull request is just to change my last name, but nothing in the actual content is changed. Sorry for the late pull request.